+++ okapi-rest/src/main/java/org/okapi/rest/metrics/payloads/SumPoint.java
@@ -14,5 +14,5 @@ import lombok.*;
 public class SumPoint {
   Long start;
   Long end;
-  long sum;
+  double sum;
 }
